+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `--force-id` flag to bypass case validation in `repo announce` - by Awiteb
+- Add `--address-file` flag to `repo announce` command - by Awiteb
+- Read the `nostr-address` file in `repo view` command - by Awiteb
 
 ## [0.1.0] - 2025-05-21